# submission/api/openapi.yaml
openapi: 3.1.0
info:
  title: LabelHub API
  version: 0.8.0
  description: Main-flow API contract for LabelHub S8 acceptance.
servers:
  - url: /api/v1
security:
  - BearerAuth: []
paths:
...
  /tasks/{taskId}/templates:
    get:
      tags: [templates]
      summary: List task template versions
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Template versions
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          templates:
                            type: array
                            items:
                              $ref: '#/components/schemas/TaskTemplate'
    post:
      tags: [templates]
      summary: Save a new task template version
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplateSchema'
      responses:
        '200':
          description: Created template
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TaskTemplate'
  /templates/{templateId}:
    get:
      tags: [templates]
      summary: Load a template version
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Template detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          template:
                            $ref: '#/components/schemas/TaskTemplate'
                          isLatest:
                            type: boolean
                          latestTemplateId:
                            type: integer
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TaskId:
      name: taskId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
    Envelope:
      type: object
      required: [data]
      properties:
        data:
          type: object
...
    TaskTemplate:
      type: object
      required: [id, taskId, version, schemaJson]
      properties:
        id:
          type: integer
        taskId:
          type: integer
        version:
          type: integer
        schemaJson:
          type: string
    TemplateSchema:
      type: object
      required: [title, fields]
      properties:
        title:
          type: string
        layout:
          type: string
          enum: [single_page]
        fields:
          type: array
          items:
            type: object
            additionalProperties: true
        export_fields:
          type: array
          items:
            type: string
      additionalProperties: true
...